split class ImportsAwareReferenceProposalCreator
